+++ tests/conftest.py
@@ -9,7 +9,7 @@ from app import create_app
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 
 
-@pytest.fixture
+@pytest.fixture(scope="session")
 def app():
     """Create and configure a new app instance for each test."""
     app = create_app()
@@ -25,7 +25,7 @@ def app():
     yield app
 
 
-@pytest.fixture
+@pytest.fixture(scope="session")
 def client(app):
     """A test client for the app."""
     return app.test_client()